Fix warns

+++ source/game_sa/Entity/Vehicle/Automobile.cpp
@@ -716,7 +716,7 @@ void CAutomobile::ProcessControl()
             if (m_nStatus == STATUS_PLAYER && !m_pHandlingData->m_bIsBus)
             {
                 if (m_nBusDoorTimerEnd) {
-                    uint32 timeStep = CTimer::GetTimeStepInMS();
+                    uint32 timeStep = (uint32)CTimer::GetTimeStepInMS();
                     if (m_nBusDoorTimerEnd <= timeStep)
                         m_nBusDoorTimerEnd = 0;
                     else
